// Qanat.Web/src/app/shared/generated/model/allocation-plan-preview-changes-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class AllocationPlanPreviewChangesDto { 
    AllocationPlanDisplayName?: string;
    ToDelete?: boolean;
    TotalPeriodsCount?: number;
    PeriodsToDeleteCount?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface AllocationPlanPreviewChangesDtoForm { 
    AllocationPlanDisplayName?: FormControl<string>;
    ToDelete?: FormControl<boolean>;
    TotalPeriodsCount?: FormControl<number>;
    PeriodsToDeleteCount?: FormControl<number>;
}

export class AllocationPlanPreviewChangesDtoFormControls { 
    public static AllocationPlanDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ToDelete = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TotalPeriodsCount = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PeriodsToDeleteCount = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}